clean: organize frontend dir
