+++ .gitignore
@@ -3,3 +3,19 @@ node_modules
 __pycache__
 .vscode
 *.csv
+
+.pnp
+.pnp.js
+
+coverage
+
+build
+
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
+
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
