+++ frontend/public/index.html
@@ -11,12 +11,12 @@
     />
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-    <link rel="stylesheet" href="header.css" />
+    <link rel="stylesheet" href="style/header.css" />
 
     <link href='https://fonts.googleapis.com/css?family=Montserrat|Cardo' rel='stylesheet' type='text/css'>
     <link href='style.css' rel='stylesheet' type='text/css'>
     <script type="text/javascript" src="https://code.jquery.com/jquery-1.9.1.js"></script>
-    <script src="header.js"></script>
+    <script src="scripts/header.js"></script>
     <title>Yuuri Kurashima</title>
   </head>
   <header class="main_h">
